WebRTC/requirements.txt
aiortc==1.10.1
openai==1.68.2
aiohttp==3.11.14
soundfile==0.13.1
langsmith==0.3.19
python-dotenv==1.1.0